moved code to function, fixed time error
I moved all the action into a file_mover function outside of the main function that is then called in main.
I also discovered that the GetLastWriteTime method was returning the folder's last modification time instead of the files inside, and found that by adding the file to source in the method call it returned the desired file modification times. 
the program seems to be working correctly now.